🔨 chore : rename folders & files from group to _class
- The term group, as used in the original conception, is broad in scope.

Related issue: #78
