+++ public/svgs/_class/index.ts
