+++ src/app/components/_class/index.ts
