+++ src/app/components/_class/modal/ClassPassword.tsx
@@ -1,57 +1,11 @@
-'use client';
-
-import React, {useState} from 'react';
-import {AxiosError} from 'axios';
 import Image from 'next/image';
-import getCheckClassSecret from '@/src/api/_class/getCheckClassSecret';
-import {ModalProps} from '@/src/interfaces/group/modal';
-import icons from '@/public/svgs/group';
+import User from '@/src/model/User';
+import logos from '@/public/images/_class';
+import {ClassPasswordProps} from '@/src/interfaces/_class/modal';
 
-const ClassJoin = ({setActiveModalId, setIsModalOpen}: ModalProps) => {
-  const isAxiosError = (error: unknown): error is AxiosError => {
-    return (error as AxiosError).isAxiosError;
-  };
-  const [inputValue, setInputValue] = useState('');
-
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setInputValue(event.target.value);
-  };
-
-  const handleSubmit = async (
-    event: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
-    event.preventDefault();
-    if (!inputValue) {
-      alert('Please input class code');
-      return;
-    }
-    try {
-      const res = await getCheckClassSecret(inputValue);
-      if (!res.secretExists) {
-        alert('This Class does not have a secret key.');
-        return;
-      }
-      setActiveModalId('');
-      if (setIsModalOpen) {
-        setIsModalOpen(true);
-      }
-    } catch (error: unknown) {
-      if (
-        isAxiosError(error) &&
-        error.response &&
-        error.response.status === 404
-      ) {
-        alert('Class code that does not exist');
-      }
-    }
-  };
-
-  const handleClose = () => {
-    setActiveModalId('');
-  };
-
+const ClassPassword = ({onClose}: ClassPasswordProps) => {
   return (
-    <div id="classJoin" className="fixed z-10 inset-0 overflow-y-auto">
+    <div id="classPassword" className="fixed z-10 inset-0 overflow-y-auto">
       <div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
         <div className="fixed inset-0 transition-opacity" aria-hidden="true">
           <div className="absolute inset-0 bg-gray-500 opacity-75"></div>
@@ -67,15 +21,15 @@ const ClassJoin = ({setActiveModalId, setIsModalOpen}: ModalProps) => {
             <div className="items-center">
               <div className="mt-3">
                 <h3 className="text-3xl leading-6 font-bold text-gray-900">
-                  Join Class
+                  Welcome! {User.name}
                 </h3>
                 <p className="mt-4 text-sm text-gray-500">
-                  Get started - it&apos;s free. No credit card needed.
+                  Please input the class password.
                 </p>
                 <div className="mt-4 flex justify-center">
                   <Image
-                    src={icons.joinIntro}
-                    alt={'joinIntro'}
+                    src={logos.github}
+                    alt={'passwordIntro'}
                     width={300}
                     height={200}
                     className="max-w-80 max-h-72 w-auto h-auto"
@@ -83,28 +37,27 @@ const ClassJoin = ({setActiveModalId, setIsModalOpen}: ModalProps) => {
                 </div>
               </div>
               <div className="mt-3">
-                <input
-                  className="mt-3 w-full inline-flex justify-center rounded-md border ring-gray-100 shadow-sm px-4 py-2 bg-white-50 text-base font-medium hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:bg-gray-50 focus:ring-gray-100 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
-                  type="text"
-                  placeholder="Input Class Code"
-                  value={inputValue}
-                  onChange={handleInputChange}
-                  required
-                />
+                <form action="">
+                  <input
+                    className="mt-3 w-full inline-flex justify-center rounded-md border ring-gray-100 shadow-sm px-4 py-2 bg-white-50 text-base font-medium hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:bg-gray-50 focus:ring-gray-100 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
+                    type="text"
+                    placeholder="Input Class Password"
+                    required
+                  />
+                </form>
               </div>
             </div>
           </div>
           <div className="flex px-4 py-3 mb-3 sm:px-6 sm:flex justify-between">
             <button
               type="button"
-              onClick={handleClose}
+              onClick={onClose}
               className="mt-3 w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-red-600 text-base font-medium text-white hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
             >
               Close
             </button>
             <button
               type="submit"
-              onClick={handleSubmit}
               className="mt-3 w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-blue-600 text-base font-medium text-white hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm"
             >
               Submit
@@ -116,4 +69,4 @@ const ClassJoin = ({setActiveModalId, setIsModalOpen}: ModalProps) => {
   );
 };
 
-export default ClassJoin;
+export default ClassPassword;
