+++ src/interfaces/_class/cardProps.ts
