+++ src/interfaces/_class/classProps.ts
