+++ src/interfaces/_class/index.ts
