+++ src/interfaces/_class/invitationProps.ts
