+++ src/interfaces/_class/inviteProps.ts
