+++ src/interfaces/_class/modal/classPasswordProps.ts
