+++ src/interfaces/_class/modal/classWaitProps.ts
