+++ src/interfaces/_class/modal/index.ts
