+++ src/interfaces/_class/modal/modalProps.ts
