+++ src/interfaces/_class/postCardProps.ts
