+++ src/interfaces/_class/roleProps.ts
