+++ src/interfaces/_class/scheduleCardProps.ts
